chore(cleanup): removes cruft in `contants.ts` and organizes them a bit
The diff is fairly large due to the variable name changes, but the tests should sniff out any copy/pasta errors.
