+++ src/actions/write/increaseVault.test.ts
@@ -1,7 +1,7 @@
 import {
   INSUFFICIENT_FUNDS_MESSAGE,
   INVALID_INPUT_MESSAGE,
-  MIN_TOKEN_LOCK_LENGTH,
+  MIN_TOKEN_LOCK_BLOCK_LENGTH,
 } from '../../constants';
 import { getBaselineState, stubbedArweaveTxId } from '../../tests/stubs';
 import { IOState } from '../../types';
@@ -17,7 +17,7 @@ describe('increaseVault', () => {
           test: {
             'existing-vault-id': {
               balance: 100,
-              end: SmartWeave.block.height + MIN_TOKEN_LOCK_LENGTH,
+              end: SmartWeave.block.height + MIN_TOKEN_LOCK_BLOCK_LENGTH,
               start: SmartWeave.block.height,
             },
           },
@@ -65,7 +65,7 @@ describe('increaseVault', () => {
         test: {
           [stubbedArweaveTxId]: {
             balance: 100,
-            end: SmartWeave.block.height + MIN_TOKEN_LOCK_LENGTH,
+            end: SmartWeave.block.height + MIN_TOKEN_LOCK_BLOCK_LENGTH,
             start: SmartWeave.block.height,
           },
         },
@@ -92,7 +92,7 @@ describe('increaseVault', () => {
         test: {
           [stubbedArweaveTxId]: {
             balance: 100,
-            end: SmartWeave.block.height + MIN_TOKEN_LOCK_LENGTH,
+            end: SmartWeave.block.height + MIN_TOKEN_LOCK_BLOCK_LENGTH,
             start: SmartWeave.block.height,
           },
         },
@@ -114,7 +114,7 @@ describe('increaseVault', () => {
         test: {
           [stubbedArweaveTxId]: {
             balance: 150,
-            end: SmartWeave.block.height + MIN_TOKEN_LOCK_LENGTH,
+            end: SmartWeave.block.height + MIN_TOKEN_LOCK_BLOCK_LENGTH,
             start: SmartWeave.block.height,
           },
         },
