+++ src/utilities.ts
@@ -1,19 +1,10 @@
 import {
-  ARNS_NAME_RESERVED_MESSAGE,
   INVALID_INPUT_MESSAGE,
-  INVALID_SHORT_NAME,
-  MAX_YEARS,
-  MINIMUM_ALLOWED_NAME_LENGTH,
-  NON_EXPIRED_ARNS_NAME_MESSAGE,
   SECONDS_IN_A_YEAR,
-  SECONDS_IN_GRACE_PERIOD,
-  SHORT_NAME_RESERVATION_UNLOCK_TIMESTAMP,
   TOTAL_IO_SUPPLY,
 } from './constants';
 import {
   ArNSAuctionData,
-  ArNSLeaseData,
-  ArNSNameData,
   Auctions,
   Balances,
   BlockHeight,
@@ -22,12 +13,7 @@ import {
   Gateway,
   Gateways,
   IOState,
-  IOToken,
-  Records,
-  RegistrationType,
   RegistryVaults,
-  ReservedNameData,
-  ReservedNames,
   VaultData,
   Vaults,
   WalletAddress,
@@ -116,54 +102,6 @@ export function resetProtocolBalance({
   };
 }
 
-export function isNameInGracePeriod({
-  currentBlockTimestamp,
-  record,
-}: {
-  currentBlockTimestamp: BlockTimestamp;
-  record: ArNSLeaseData;
-}): boolean {
-  if (!record.endTimestamp) return false;
-  const recordIsExpired = currentBlockTimestamp.valueOf() > record.endTimestamp;
-  return (
-    recordIsExpired &&
-    record.endTimestamp + SECONDS_IN_GRACE_PERIOD >
-      currentBlockTimestamp.valueOf()
-  );
-}
-
-export function getMaxAllowedYearsExtensionForRecord({
-  currentBlockTimestamp,
-  record,
-}: {
-  currentBlockTimestamp: BlockTimestamp;
-  record: ArNSLeaseData;
-}): number {
-  if (!record.endTimestamp) {
-    return 0;
-  }
-  // if expired return 0 because it cannot be extended and must be re-bought
-  if (
-    currentBlockTimestamp.valueOf() >
-    record.endTimestamp + SECONDS_IN_GRACE_PERIOD
-  ) {
-    return 0;
-  }
-
-  if (isNameInGracePeriod({ currentBlockTimestamp, record })) {
-    return MAX_YEARS;
-  }
-
-  // TODO: should we put this as the ceiling? or should we allow people to extend as soon as it is purchased
-  const yearsRemainingOnLease = Math.ceil(
-    (record.endTimestamp.valueOf() - currentBlockTimestamp.valueOf()) /
-      SECONDS_IN_A_YEAR,
-  );
-
-  // a number between 0 and 5 (MAX_YEARS)
-  return MAX_YEARS - yearsRemainingOnLease;
-}
-
 export function getInvalidAjvMessage(
   validator: any,
   input: any,
@@ -176,159 +114,6 @@ export function getInvalidAjvMessage(
       return `${key} ('${value}') ${e.message}`;
     })
     .join(', ')}`;
-}
-
-export function isExistingActiveRecord({
-  record,
-  currentBlockTimestamp,
-}: {
-  record: ArNSNameData | undefined;
-  currentBlockTimestamp: BlockTimestamp;
-}): boolean {
-  if (!record) return false;
-
-  if (record.type === 'permabuy') {
-    return true;
-  }
-
-  if (record.type === 'lease' && record.endTimestamp) {
-    return (
-      record.endTimestamp > currentBlockTimestamp.valueOf() ||
-      isNameInGracePeriod({ currentBlockTimestamp, record })
-    );
-  }
-  return false;
-}
-
-export function isShortNameRestricted({
-  name,
-  currentBlockTimestamp,
-}: {
-  name: string;
-  currentBlockTimestamp: BlockTimestamp;
-}): boolean {
-  return (
-    name.length < MINIMUM_ALLOWED_NAME_LENGTH &&
-    currentBlockTimestamp.valueOf() < SHORT_NAME_RESERVATION_UNLOCK_TIMESTAMP
-  );
-}
-
-export function isActiveReservedName({
-  caller,
-  reservedName,
-  currentBlockTimestamp,
-}: {
-  caller: string | undefined;
-  reservedName: ReservedNameData | undefined;
-  currentBlockTimestamp: BlockTimestamp;
-}): boolean {
-  if (!reservedName) return false;
-  const target = reservedName.target;
-  const endTimestamp = reservedName.endTimestamp;
-  const permanentlyReserved = !target && !endTimestamp;
-  if (permanentlyReserved) {
-    return true;
-  }
-  const callerNotTarget = !caller || target !== caller;
-  const notExpired =
-    endTimestamp && endTimestamp > currentBlockTimestamp.valueOf();
-  if (callerNotTarget && notExpired) {
-    return true;
-  }
-  return false;
-}
-
-export function isNameAvailableForAuction({
-  name,
-  record,
-  reservedName,
-  caller,
-  currentBlockTimestamp,
-}: {
-  name: string;
-  record: ArNSNameData | undefined;
-  caller: string;
-  reservedName: ReservedNameData | undefined;
-  currentBlockTimestamp: BlockTimestamp;
-}): boolean {
-  return (
-    !isExistingActiveRecord({ record, currentBlockTimestamp }) &&
-    !isActiveReservedName({ reservedName, caller, currentBlockTimestamp }) &&
-    !isShortNameRestricted({ name, currentBlockTimestamp })
-  );
-}
-
-export function isNameRequiredToBeAuction({
-  name,
-  type,
-}: {
-  name: string;
-  type: RegistrationType;
-}): boolean {
-  return type === 'permabuy' && name.length < 12;
-}
-
-export function assertAvailableRecord({
-  caller,
-  name,
-  records,
-  reserved,
-  currentBlockTimestamp,
-}: {
-  caller: string | undefined; // TODO: type for this
-  name: DeepReadonly<string>;
-  records: DeepReadonly<Records>;
-  reserved: DeepReadonly<ReservedNames>;
-  currentBlockTimestamp: BlockTimestamp;
-}): void {
-  if (
-    isExistingActiveRecord({
-      record: records[name],
-      currentBlockTimestamp,
-    })
-  ) {
-    throw new ContractError(NON_EXPIRED_ARNS_NAME_MESSAGE);
-  }
-  if (
-    isActiveReservedName({
-      caller,
-      reservedName: reserved[name],
-      currentBlockTimestamp,
-    })
-  ) {
-    throw new ContractError(ARNS_NAME_RESERVED_MESSAGE);
-  }
-
-  if (isShortNameRestricted({ name, currentBlockTimestamp })) {
-    throw new ContractError(INVALID_SHORT_NAME);
-  }
-}
-
-export function calculateExistingAuctionBidForCaller({
-  caller,
-  auction,
-  submittedBid,
-  requiredMinimumBid,
-}: {
-  caller: string;
-  auction: ArNSAuctionData;
-  submittedBid: number | undefined; // TODO: change to IOToken
-  requiredMinimumBid: IOToken; // TODO: change to IOToken
-}): IOToken {
-  if (submittedBid && submittedBid < requiredMinimumBid.valueOf()) {
-    throw new ContractError(
-      `The bid (${submittedBid} IO) is less than the current required minimum bid of ${requiredMinimumBid.valueOf()} IO.`,
-    );
-  }
-
-  let finalBid = submittedBid
-    ? Math.min(submittedBid, requiredMinimumBid.valueOf())
-    : requiredMinimumBid.valueOf();
-
-  if (caller === auction.initiator) {
-    finalBid -= auction.floorPrice;
-  }
-  return new IOToken(finalBid);
 }
 
 export function isGatewayJoined({
@@ -412,8 +197,4 @@ export function incrementBalance(
   } else {
     balances[address] = amount;
   }
-}
-
-export function isLeaseRecord(record: ArNSNameData): record is ArNSLeaseData {
-  return record.type === 'lease';
 }
